sdks/python/apache_beam/examples/preprocessing/kfp/preprocessing/requirements.txt
apache_beam==2.40.0
requests==2.28.1
torch==1.12.0
torchvision==0.13.0
numpy==1.22.4
Pillow==9.2.0